+++ drools-platform-manager/drools-platform-manager-ui/src/main/java/org/chtijbug/drools/platform/backend/wsclient/listener/IsAliveListener.java
@@ -17,5 +17,5 @@ package org.chtijbug.drools.platform.backend.wsclient.listener;
 
 
 public interface IsAliveListener {
-    public void messageReceived(Integer ruleBaseID);
+    public void messageReceived(Long ruleBaseID);
 }
